requirements_eval.txt
# Core dependencies for evaluation
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
numpy>=1.21.0
torch>=1.10.0
pyyaml>=6.0
tqdm>=4.62.0

# Optional for enhanced plotting
plotly>=5.0.0
kaleido>=0.2.1
